Add tests for oc adm compare
This commit adds tests for the `oc adm compare` command. The test cases include comparing CRs from both a live cluster and a local directory. These tests run `oc adm compare` with various outputs and verify that the expected stdout and stderr are obtained. A `--update` flag has been introduced to enable easy test writing and modification. The tests execute the command with a concurrency value greater than 1, enabling race detection when running the tests with the `-race` flag.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>