Add option to pull reference config from a URL
In this commit, the option to pass a URL with the -r flag containing a raw
version of the reference config was introduced. To support this change, HTTPFS
was introduced, implementing a file system interface for HTTP servers
returning raw file content as part of the HTTP response body. Additionally,
the template.ParseFiles function was replaced with the template.ParseFS function,
allowing retrieval of the reference from a local file system or an HTTP file
system. Additional tests were added to support this case.

Furthermore, the change included calling parseFS each time with one file
instead of multiple files, enabling the reporting of all template parsing
errors to the user in one run. This also 
fixed problems with templates
included in subdirectories, which were reported missing in the summary even if
they were not.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>