Improve live cluster search by using GroupVersionKind
This update modifies how resources are found in live cluster mode. Previously,
we searched for resources based only on their  kind, rather
than their GroupVersionKind. Now, 
the code is adjusted to search for each resource kind found in the reference
CR across all API group-versions where that kind exists. To enable this
change, a list of group-versions is generated for each supported kind. This
resolves issues where the same type was returned by multiple API resources,
causing the visitor to miss some resources of that type.

Signed-off-by: Alina Sudakov <asudakov@redhat.com>